Moving error definitions back to configuration package
