+++ rosetta/fail/invalid_currency.go
@@ -14,6 +14,10 @@
 
 package fail
 
+import (
+	"github.com/optakt/flow-dps/rosetta/configuration"
+)
+
 // InvalidCurrency error is returned when the currency identifier is invalid.
 // Currency identifier is considered invalid when the specified number of decimals
 // is different from the configured number of decimals.
@@ -29,5 +33,5 @@ func (i InvalidCurrency) Error() string {
 
 // RosettaError returns the error information in a Rosetta-compatible format.
 func (i InvalidCurrency) RosettaError() Error {
-	return newError(ErrorInvalidCurrency, i.Description, i.Details...)
+	return newError(configuration.ErrorInvalidCurrency, i.Description, i.Details...)
 }
